updated autofocus

+++ src/app/api/attendance/route.ts
@@ -11,13 +11,13 @@ export async function POST(req: Request) {
 
   if (!status) {
     await sql(
-      "DELETE FROM entries WHERE course_id = $1 AND date = coalesce($2, (CURRENT_DATE at time zone 'Africa/Cairo')::date)",
-      [course_id, date],
+      "DELETE FROM entries USING courses WHERE courses.id = course_times.course_id AND WHERE course_id = $1 AND date = coalesce($2, (CURRENT_DATE at time zone 'Africa/Cairo')::date) AND user_id = $3",
+      [course_id, date, user.id],
     );
   } else {
     await sql(
-      "INSERT INTO entries (course_id, type, date) VALUES ($1, $2, coalesce($3, (CURRENT_DATE at time zone 'Africa/Cairo'))) ON CONFLICT (course_id, date) DO UPDATE SET type = $2",
-      [course_id, status, date],
+      "INSERT INTO entries ((SELECT id FROM courses WHERE id = $1 AND user_id = $4), type, date) VALUES ($1, $2, coalesce($3, (CURRENT_DATE at time zone 'Africa/Cairo'))) ON CONFLICT (course_id, date) DO UPDATE SET type = $2",
+      [course_id, status, date, user.id],
     );
   }
